Remove undefined as option from optional props. Adjust Main and MainSection ontology.
