+++ .storybook/main.js
@@ -9,6 +9,7 @@ module.exports = {
     reactDocgen: 'react-docgen-typescript',
     reactDocgenTypescriptOptions: {
       shouldExtractLiteralValuesFromEnum: true,
+      shouldRemoveUndefinedFromOptional: true,
       tsconfigPath: tsconfig,
     },
   },
