+++ src/view/View.tsx
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { MainSection } from '../surfaces/MainSection'
+import { Main } from '../surfaces/Main'
 import { FluentKitProvider } from '../lib/FluentKitProvider'
 import { sectionProps } from '../blocks/Section'
 import { theme } from '../lib/theme'
@@ -9,7 +9,7 @@ export const viewProps = z.object({
   sidebar: z.object({}).optional(),
   toolbar: z.object({}).optional(),
   modal: z.object({}).optional(),
-  mainSection: sectionProps,
+  main: sectionProps,
   theme: theme.optional(),
   dir: dir.optional(),
 })
@@ -18,10 +18,10 @@ export type ViewProps = z.infer<typeof viewProps>
 
 /** An experience provided to the user via their device’s canvas. */
 export const View = (props: ViewProps) => {
-  const { mainSection, theme = 'light', dir = 'ltr' } = viewProps.parse(props)
+  const { main, theme = 'light', dir = 'ltr' } = viewProps.parse(props)
   return (
     <FluentKitProvider {...{ theme, dir }}>
-      <MainSection {...mainSection} />
+      <Main {...main} />
     </FluentKitProvider>
   )
 }
